refactor: lazy load large screens
