+++ app/store/[storeId]/admin/products.tsx
@@ -1,123 +1,12 @@
-import React, { useEffect, useState } from 'react';
-import { View, Text, StyleSheet, FlatList, TouchableOpacity } from 'react-native';
-import { useLocalSearchParams } from 'expo-router';
-import { useTheme } from '../../../../contexts/ThemeContext';
-import chain from '../../../../services/chain';
-import { Product } from '../../../../types';
+import React, { Suspense } from 'react';
+import Spinner from '../../../../components/ui/Spinner';
 
-let listProducts: (() => Promise<Product[]>) | undefined;
-if (chain === 'ton') {
-  ({ listProducts } = require('@/features/products/services/tonProducts'));
-}
-import ProductCard from '@/features/products/components/ProductCard';
-import ProductFormModal from '@/features/products/components/ProductFormModal';
-import { useAccountId } from '@/features/auth/services/nearAuth';
-import { useAuth } from '@/features/auth/AuthContext';
+const ProductsScreen = React.lazy(() => import('./_ProductsScreen'));
 
-const ITEM_HEIGHT = 200;
-
-export default function StoreProductsScreen() {
-  const { storeId } = useLocalSearchParams<{ storeId: string }>();
-  const { colors } = useTheme();
-  const [products, setProducts] = useState<Product[]>([]);
-  const [formVisible, setFormVisible] = useState(false);
-  const [editingProduct, setEditingProduct] = useState<Product | null>(null);
-  const address = useAccountId();
-  const { isStoreOwner } = useAuth();
-
-  useEffect(() => {
-    const load = async () => {
-      if (!storeId || !isStoreOwner || address !== storeId || !listProducts) return;
-      const all = await listProducts();
-      setProducts(all.filter((p) => p.storeId === storeId));
-    };
-    load();
-  }, [storeId, address, isStoreOwner]);
-
-  const openForm = (p?: Product) => {
-    setEditingProduct(p || null);
-    setFormVisible(true);
-  };
-
-  const handleSaved = (p: Product, isNew: boolean) => {
-    if (isNew) {
-      setProducts((prev) => [...prev, p]);
-    } else {
-      setProducts((prev) => prev.map((prod) => (prod.id === p.id ? p : prod)));
-    }
-  };
-
-  const handleDeleted = (productId: string) => {
-    setProducts((prev) => prev.filter((p) => p.id !== productId));
-  };
-
-  if (!isStoreOwner || address !== storeId) {
-    return (
-      <View
-        style={[
-          styles.container,
-          {
-            backgroundColor: colors.background,
-            justifyContent: 'center',
-            alignItems: 'center',
-          },
-        ]}
-      >
-        <Text style={{ color: colors.text.primary }}>יש להתחבר לארנק המתאים</Text>
-      </View>
-    );
-  }
-
+export default function ProductsRoute(props: any) {
   return (
-    <View style={[styles.container, { backgroundColor: colors.background }]}> 
-      <TouchableOpacity
-        style={[styles.addButton, { borderColor: colors.border.primary }]}
-        onPress={() => openForm()}
-      >
-        <Text style={{ color: colors.text.primary }}>הוסף מוצר</Text>
-      </TouchableOpacity>
-      <FlatList
-        data={products}
-        keyExtractor={(p) => p.id}
-        renderItem={({ item: p }) => (
-          <TouchableOpacity style={{ marginBottom: 12 }}>
-            <ProductCard
-              product={p}
-              isOwner
-              onEdit={() => openForm(p)}
-              onDelete={(id) => handleDeleted(id)}
-            />
-          </TouchableOpacity>
-        )}
-        ListEmptyComponent={
-          <Text style={{ color: colors.text.secondary, textAlign: 'center' }}>אין מוצרים</Text>
-        }
-        contentContainerStyle={styles.list}
-        getItemLayout={(_, index) => ({ length: ITEM_HEIGHT, offset: ITEM_HEIGHT * index, index })}
-      />
-      <ProductFormModal
-        visible={formVisible}
-        product={editingProduct}
-        onClose={() => setFormVisible(false)}
-        onSaved={handleSaved}
-        onDeleted={handleDeleted}
-      />
-    </View>
+    <Suspense fallback={<Spinner />}>
+      <ProductsScreen {...props} />
+    </Suspense>
   );
 }
-
-const styles = StyleSheet.create({
-  container: { flex: 1, padding: 16 },
-  addButton: {
-    paddingVertical: 12,
-    paddingHorizontal: 16,
-    borderWidth: 1,
-    borderRadius: 8,
-    alignItems: 'center',
-    marginBottom: 16,
-  },
-  list: {
-    paddingBottom: 24,
-  },
-});
-
